[refactor/#179] Member Controller test 재구현

+++ module-auth/src/main/java/com/inhabas/api/auth/domain/oauth2/member/domain/service/MemberService.java
@@ -28,7 +28,7 @@ public interface MemberService {
 
     List<ApprovedMemberManagementDto> getApprovedMembersBySearchAndRole(String search);
 
-    void updateUnapprovedMembers(List<Integer> memberIdList, String state);
+    void updateUnapprovedMembers(List<Long> memberIdList, String state);
 
     void updateApprovedMembers(List<Long> memberIdList, Role role);
